+++ src/core/ReactRef.js
@@ -11,6 +11,7 @@
 
 'use strict';
 
+var ReactOwner = require('ReactOwner');
 var ReactUpdates = require('ReactUpdates');
 
 var accumulate = require('accumulate');
@@ -81,15 +82,70 @@ assign(ReactRef.prototype, {
   }
 });
 
-ReactRef.attachRef = function(ref, value) {
+function attachFirstClassRef(ref, value) {
   ref._value = value.getPublicInstance();
-};
+}
 
-ReactRef.detachRef = function(ref, value) {
+function detachFirstClassRef(ref, value) {
   // Check that `component` is still the current ref because we do not want to
   // detach the ref if another component stole it.
   if (ref._value === value) {
     ref._value = null;
+  }
+}
+
+function attachRef(ref, component, owner) {
+  if (ref instanceof ReactRef) {
+    attachFirstClassRef(ref, component);
+  } else {
+    ReactOwner.addComponentAsRefTo(component, ref, owner);
+  }
+}
+
+function detachRef(ref, component, owner) {
+  if (ref instanceof ReactRef) {
+    detachFirstClassRef(ref, component);
+  } else {
+    ReactOwner.removeComponentAsRefFrom(component, ref, owner);
+  }
+}
+
+ReactRef.attachRefs = function(instance, element) {
+  var ref = element.ref;
+  if (ref != null) {
+    attachRef(ref, instance, element._owner);
+  }
+};
+
+ReactRef.updateRefs = function(instance, prevElement, nextElement) {
+  // If either the owner or a `ref` has changed, make sure the newest owner
+  // has stored a reference to `this`, and the previous owner (if different)
+  // has forgotten the reference to `this`. We use the element instead
+  // of the public this.props because the post processing cannot determine
+  // a ref. The ref conceptually lives on the element.
+
+  // TODO: Should this even be possible? The owner cannot change because
+  // it's forbidden by shouldUpdateReactComponent. The ref can change
+  // if you swap the keys of but not the refs. Reconsider where this check
+  // is made. It probably belongs where the key checking and
+  // instantiateReactComponent is done.
+
+  if (nextElement._owner !== prevElement._owner ||
+      nextElement.ref !== prevElement.ref) {
+    if (prevElement.ref != null) {
+      detachRef(prevElement.ref, instance, prevElement._owner);
+    }
+    // Correct, even if the owner is the same, and only the ref has changed.
+    if (nextElement.ref != null) {
+      attachRef(nextElement.ref, instance, nextElement._owner);
+    }
+  }
+};
+
+ReactRef.detachRefs = function(instance, element) {
+  var ref = element.ref;
+  if (ref != null) {
+    detachRef(ref, instance, element._owner);
   }
 };
